# Refactor state and contract modules