--- a/cmd/juno/dbcmd.go
+++ b/cmd/juno/dbcmd.go
@@ -206,7 +206,7 @@ func dbSize(cmd *cobra.Command, args []string) error {
 		totalSize += bucketItem.Size
 		totalCount += bucketItem.Count
 
-		if utils.AnyOf(b, db.StateTrie, db.ContractStorage, db.Class, db.ContractNonce, db.ContractDeploymentHeight) {
+		if utils.AnyOf(b, db.StateTrie, db.ContractStorage, db.Class, db.Contract) {
 			withoutHistorySize += bucketItem.Size
 			withHistorySize += bucketItem.Size