--- a/core/contract.go
+++ b/core/contract.go
@@ -7,6 +7,7 @@ import (
 	"github.com/NethermindEth/juno/core/felt"
 	"github.com/NethermindEth/juno/core/trie"
 	"github.com/NethermindEth/juno/db"
+	"github.com/NethermindEth/juno/encoder"
 )
 
 // contract storage has fixed height at 251
@@ -17,51 +18,141 @@ var (
 	ErrContractAlreadyDeployed = errors.New("contract already deployed")
 )
 
-// NewContractUpdater creates an updater for the contract instance at the given address.
-// Deploy should be called for contracts that were just deployed to the network.
-func NewContractUpdater(addr *felt.Felt, txn db.Transaction) (*ContractUpdater, error) {
-	contractDeployed, err := deployed(addr, txn)
-	if err != nil {
-		return nil, err
-	}
+type OnValueChanged = func(location, oldValue *felt.Felt) error
 
-	if !contractDeployed {
-		return nil, ErrContractNotDeployed
-	}
-
-	return &ContractUpdater{
-		Address: addr,
-		txn:     txn,
-	}, nil
+type StateContract struct {
+	// ClassHash is the hash of the contract's class
+	ClassHash *felt.Felt
+	// Nonce is the contract's nonce
+	Nonce *felt.Felt
+	// DeployHeight is the height at which the contract is deployed
+	DeployHeight uint64
+	// Address that this contract instance is deployed to
+	Address *felt.Felt `cbor:"-"`
+	// Storage is the contract's storage
+	Storage map[felt.Felt]*felt.Felt `cbor:"-"`
 }
 
-// DeployContract sets up the database for a new contract.
-func DeployContract(addr, classHash *felt.Felt, txn db.Transaction) (*ContractUpdater, error) {
-	contractDeployed, err := deployed(addr, txn)
+func NewStateContract(
+	addr *felt.Felt,
+	classHash *felt.Felt,
+	nonce *felt.Felt,
+	storage map[felt.Felt]*felt.Felt,
+	DeployHeight uint64,
+) *StateContract {
+	sc := &StateContract{
+		Address:      addr,
+		ClassHash:    classHash,
+		Nonce:        nonce,
+		Storage:      storage,
+		DeployHeight: DeployHeight,
+	}
+
+	if storage == nil {
+		sc.Storage = make(map[felt.Felt]*felt.Felt)
+	}
+
+	return sc
+}
+
+func (c *StateContract) StorageRoot(txn db.Transaction) (*felt.Felt, error) {
+	storageTrie, err := storage(c.Address, txn)
 	if err != nil {
 		return nil, err
 	}
 
-	if contractDeployed {
-		return nil, ErrContractAlreadyDeployed
+	return storageTrie.Root()
+}
+
+func (c *StateContract) GetStorage(key *felt.Felt, txn db.Transaction) (*felt.Felt, error) {
+	if c.Storage != nil {
+		if val, ok := c.Storage[*key]; ok {
+			return val, nil
+		}
 	}
 
-	err = setClassHash(txn, addr, classHash)
+	// get from db
+	storage, err := storage(c.Address, txn)
 	if err != nil {
 		return nil, err
 	}
 
-	c, err := NewContractUpdater(addr, txn)
+	return storage.Get(key)
+}
+
+func (c *StateContract) Commit(txn db.Transaction, cb OnValueChanged) error {
+	storageTrie, err := storage(c.Address, txn)
 	if err != nil {
+		return err
+	}
+
+	for key, value := range c.Storage {
+		oldVal, err := storageTrie.Put(&key, value)
+		if err != nil {
+			return err
+		}
+
+		if oldVal != nil {
+			if err = cb(&key, oldVal); err != nil {
+				return err
+			}
+		}
+	}
+
+	if err := storageTrie.Commit(); err != nil {
+		return err
+	}
+
+	contractBytes, err := encoder.Marshal(c)
+	if err != nil {
+		return err
+	}
+
+	return txn.Set(db.Contract.Key(c.Address.Marshal()), contractBytes)
+}
+
+// Purge eliminates the contract instance, deleting all associated data from database
+// assumes storage is cleared in revert process
+func (c *StateContract) Purge(txn db.Transaction) error {
+	addrBytes := c.Address.Marshal()
+
+	if err := txn.Delete(db.Contract.Key(addrBytes)); err != nil {
+		return err
+	}
+
+	return txn.Delete(db.ContractDeploymentHeight.Key(addrBytes))
+}
+
+// GetContract is a wrapper around getContract which checks if a contract is deployed
+func GetContract(addr *felt.Felt, txn db.Transaction) (*StateContract, error) {
+	contract, err := getContract(addr, txn)
+	if err != nil {
+		if errors.Is(err, db.ErrKeyNotFound) {
+			return nil, ErrContractNotDeployed
+		}
 		return nil, err
 	}
 
-	err = c.UpdateNonce(&felt.Zero)
-	if err != nil {
+	return contract, nil
+}
+
+// getContract gets a contract instance from the database.
+func getContract(addr *felt.Felt, txn db.Transaction) (*StateContract, error) {
+	key := db.Contract.Key(addr.Marshal())
+	var contract StateContract
+	if err := txn.Get(key, func(val []byte) error {
+		if err := encoder.Unmarshal(val, &contract); err != nil {
+			return err
+		}
+
+		contract.Address = addr
+		contract.Storage = make(map[felt.Felt]*felt.Felt)
+
+		return nil
+	}); err != nil {
 		return nil, err
 	}
-
-	return c, nil
+	return &contract, nil
 }
 
 // ContractAddress computes the address of a Starknet contract.
@@ -79,129 +170,9 @@ func ContractAddress(callerAddress, classHash, salt *felt.Felt, constructorCallD
 	)
 }
 
-func deployed(addr *felt.Felt, txn db.Transaction) (bool, error) {
-	_, err := ContractClassHash(addr, txn)
-	if errors.Is(err, db.ErrKeyNotFound) {
-		return false, nil
-	}
-	if err != nil {
-		return false, err
-	}
-	return true, nil
-}
-
-// ContractUpdater is a helper to update an existing contract instance.
-type ContractUpdater struct {
-	// Address that this contract instance is deployed to
-	Address *felt.Felt
-	// txn to access the database
-	txn db.Transaction
-}
-
-// Purge eliminates the contract instance, deleting all associated data from storage
-// assumes storage is cleared in revert process
-func (c *ContractUpdater) Purge() error {
-	addrBytes := c.Address.Marshal()
-	buckets := []db.Bucket{db.ContractNonce, db.ContractClassHash}
-
-	for _, bucket := range buckets {
-		if err := c.txn.Delete(bucket.Key(addrBytes)); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-// ContractNonce returns the amount transactions sent from this contract.
-// Only account contracts can have a non-zero nonce.
-func ContractNonce(addr *felt.Felt, txn db.Transaction) (*felt.Felt, error) {
-	key := db.ContractNonce.Key(addr.Marshal())
-	var nonce *felt.Felt
-	if err := txn.Get(key, func(val []byte) error {
-		nonce = new(felt.Felt)
-		nonce.SetBytes(val)
-		return nil
-	}); err != nil {
-		return nil, err
-	}
-	return nonce, nil
-}
-
-// UpdateNonce updates the nonce value in the database.
-func (c *ContractUpdater) UpdateNonce(nonce *felt.Felt) error {
-	nonceKey := db.ContractNonce.Key(c.Address.Marshal())
-	return c.txn.Set(nonceKey, nonce.Marshal())
-}
-
-// ContractRoot returns the root of the contract storage.
-func ContractRoot(addr *felt.Felt, txn db.Transaction) (*felt.Felt, error) {
-	cStorage, err := storage(addr, txn)
-	if err != nil {
-		return nil, err
-	}
-	return cStorage.Root()
-}
-
-type OnValueChanged = func(location, oldValue *felt.Felt) error
-
-// UpdateStorage applies a change-set to the contract storage.
-func (c *ContractUpdater) UpdateStorage(diff map[felt.Felt]*felt.Felt, cb OnValueChanged) error {
-	cStorage, err := storage(c.Address, c.txn)
-	if err != nil {
-		return err
-	}
-	// apply the diff
-	for key, value := range diff {
-		oldValue, pErr := cStorage.Put(&key, value)
-		if pErr != nil {
-			return pErr
-		}
-
-		if oldValue != nil {
-			if err = cb(&key, oldValue); err != nil {
-				return err
-			}
-		}
-	}
-
-	return cStorage.Commit()
-}
-
-func ContractStorage(addr, key *felt.Felt, txn db.Transaction) (*felt.Felt, error) {
-	cStorage, err := storage(addr, txn)
-	if err != nil {
-		return nil, err
-	}
-	return cStorage.Get(key)
-}
-
-// ContractClassHash returns hash of the class that the contract at the given address instantiates.
-func ContractClassHash(addr *felt.Felt, txn db.Transaction) (*felt.Felt, error) {
-	key := db.ContractClassHash.Key(addr.Marshal())
-	var classHash *felt.Felt
-	if err := txn.Get(key, func(val []byte) error {
-		classHash = new(felt.Felt)
-		classHash.SetBytes(val)
-		return nil
-	}); err != nil {
-		return nil, err
-	}
-	return classHash, nil
-}
-
-func setClassHash(txn db.Transaction, addr, classHash *felt.Felt) error {
-	classHashKey := db.ContractClassHash.Key(addr.Marshal())
-	return txn.Set(classHashKey, classHash.Marshal())
-}
-
-// Replace replaces the class that the contract instantiates
-func (c *ContractUpdater) Replace(classHash *felt.Felt) error {
-	return setClassHash(c.txn, c.Address, classHash)
-}
-
 // storage returns the [core.Trie] that represents the
 // storage of the contract.
+// TODO(weiihann): how to deal with the root key?
 func storage(addr *felt.Felt, txn db.Transaction) (*trie.Trie, error) {
 	addrBytes := addr.Marshal()
 	trieTxn := trie.NewStorage(txn, db.ContractStorage.Key(addrBytes))